Extracted Class_ methods to traits for use with Trait_
